bushido:hooks now creates both user and app hooks
